requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
numpy==1.21.2
opencv-python==4.5.3.56
face-recognition==1.3.0
gunicorn==20.1.0
python-dotenv==0.19.0